Separate components on list of entities
