+++ frontend/src/components/entity/EntityList.test.js
@@ -1,0 +1,12 @@
+/**
+ * @jest-environment jsdom
+ */
+
+import React from "react";
+import { render } from "@testing-library/react";
+
+import EntityList from "./EntityList";
+
+test("should render a component with essential props", function () {
+  expect(() => render(<EntityList entities={[]} />)).not.toThrow();
+});
