+++ frontend/src/pages/Entity.js
@@ -1,31 +1,16 @@
-import React, { useEffect, useState } from "react";
-import Button from "@material-ui/core/Button";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TablePagination,
-  TableRow,
-} from "@material-ui/core";
-import Paper from "@material-ui/core/Paper";
+import React from "react";
 import { Link } from "react-router-dom";
-import HistoryIcon from "@material-ui/icons/History";
-import GroupIcon from "@material-ui/icons/Group";
-import { makeStyles } from "@material-ui/core/styles";
 import Typography from "@material-ui/core/Typography";
 import AironeBreadcrumbs from "../components/common/AironeBreadcrumbs";
-import { deleteEntity, getEntities } from "../utils/AironeAPIClient";
-import EditButton from "../components/common/EditButton";
+import { getEntities } from "../utils/AironeAPIClient";
+import { useAsync } from "react-use";
+import EntityList from "../components/entity/EntityList";
 import CreateButton from "../components/common/CreateButton";
-import DeleteButton from "../components/common/DeleteButton";
+import Button from "@material-ui/core/Button";
+import { makeStyles } from "@material-ui/core/styles";
 
 const useStyles = makeStyles((theme) => ({
   button: {
-    margin: theme.spacing(1),
-  },
-  entityName: {
     margin: theme.spacing(1),
   },
 }));
@@ -33,53 +18,11 @@ const useStyles = makeStyles((theme) => ({
 export default function Entity({}) {
   const classes = useStyles();
 
-  const [keyword, setKeyword] = useState("");
-  const [entities, setEntities] = useState([]);
-  const [filteredEntities, setFilteredEntities] = useState([]);
-
-  const [page, setPage] = React.useState(0);
-  const [rowsPerPage, setRowsPerPage] = React.useState(100);
-
-  const [updated, setUpdated] = useState(false);
-
-  useEffect(() => {
-    getEntities()
+  const entities = useAsync(async () => {
+    return getEntities()
       .then((resp) => resp.json())
-      .then((data) => {
-        setEntities(data.entities);
-        setFilteredEntities(data.entities);
-      });
-    setUpdated(false);
-  }, [updated]);
-
-  const onChangeKeyword = (event) => {
-    setKeyword(event.target.value);
-  };
-
-  const onKeyPressKeyword = (event) => {
-    if (event.key === "Enter" && keyword) {
-      setFilteredEntities(
-        entities.filter((entity) => {
-          return entity.name.indexOf(keyword) !== -1;
-        })
-      );
-    } else {
-      setFilteredEntities(entities);
-    }
-  };
-
-  const handleChangePage = (event, newPage) => {
-    setPage(newPage);
-  };
-
-  const handleChangeRowsPerPage = (event) => {
-    setRowsPerPage(+event.target.value);
-    setPage(0);
-  };
-
-  const handleDelete = (event, entityId) => {
-    deleteEntity(entityId).then((_) => setUpdated(true));
-  };
+      .then((data) => data.entities);
+  });
 
   return (
     <div className="container-fluid">
@@ -113,93 +56,11 @@ export default function Entity({}) {
               インポート
             </Button>
           </div>
-          <div className="float-right"></div>
+          <div className="float-right" />
         </div>
       </div>
 
-      <Paper>
-        <TableContainer component={Paper}>
-          <Table>
-            <TableHead>
-              <TableRow>
-                <TableCell>
-                  <span className={classes.entityName}>エンティティ名</span>
-                  <input
-                    className={classes.entityName}
-                    text="text"
-                    placeholder="絞り込む"
-                    value={keyword}
-                    onChange={onChangeKeyword}
-                    onKeyPress={onKeyPressKeyword}
-                  />
-                </TableCell>
-                <TableCell>
-                  <Typography>備考</Typography>
-                </TableCell>
-                <TableCell align="right" />
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {filteredEntities
-                .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-                .map((entity) => (
-                  <TableRow>
-                    <TableCell>
-                      <Typography
-                        component={Link}
-                        to={`/new-ui/entities/${entity.id}/entries`}
-                      >
-                        {entity.name}
-                      </Typography>
-                    </TableCell>
-                    <TableCell>
-                      <Typography>{entity.note}</Typography>
-                    </TableCell>
-                    <TableCell align="right">
-                      <EditButton to={`/new-ui/entities/${entity.id}`}>
-                        エンティティ編集
-                      </EditButton>
-                      <Button
-                        variant="contained"
-                        color="primary"
-                        className={classes.button}
-                        startIcon={<HistoryIcon />}
-                        component={Link}
-                        to={`/new-ui/entities/${entity.id}/history`}
-                      >
-                        変更履歴
-                      </Button>
-                      <Button
-                        variant="contained"
-                        color="primary"
-                        className={classes.button}
-                        startIcon={<GroupIcon />}
-                        component={Link}
-                        to={`/new-ui/acl/${entity.id}`}
-                      >
-                        ACL
-                      </Button>
-                      <DeleteButton
-                        onConfirmed={(e) => handleDelete(e, entity.id)}
-                      >
-                        削除
-                      </DeleteButton>
-                    </TableCell>
-                  </TableRow>
-                ))}
-            </TableBody>
-          </Table>
-        </TableContainer>
-        <TablePagination
-          rowsPerPageOptions={[100, 250, 1000]}
-          component="div"
-          count={filteredEntities.length}
-          rowsPerPage={rowsPerPage}
-          page={page}
-          onChangePage={handleChangePage}
-          onChangeRowsPerPage={handleChangeRowsPerPage}
-        />
-      </Paper>
+      {!entities.loading && <EntityList entities={entities.value} />}
     </div>
   );
 }
